+++ frontend/src/components/CardBoard/CardBody/CardEnd.tsx
@@ -38,13 +38,13 @@ const CardEnd: React.FC<CardEndProps> = React.memo(
 
     const boardTypeCaption = useMemo(() => {
       if (isSubBoard && !isDashboard) return 'Responsible';
-      if (isSubBoard && !team && isDashboard) return 'Team';
+      if (isSubBoard && isDashboard) return 'Team';
       if (team) return 'Team';
       return 'Personal';
     }, [isDashboard, isSubBoard, team]);
 
     const boardOwnerName = useMemo(() => {
-      if (team) {
+      if (team && !isSubBoard) {
         return team?.name;
       }
       if (isSubBoard && isDashboard && index !== undefined) {
@@ -67,7 +67,7 @@ const CardEnd: React.FC<CardEndProps> = React.memo(
             {boardOwnerName}
           </Text>
           <CardAvatars
-            listUsers={!team ? users : team.users}
+            listUsers={!team || isSubBoard ? users : team.users}
             responsible={false}
             teamAdmins={false}
             userId={userId}