+++ frontend/src/utils/verifyIfIsNewJoiner.ts
@@ -1,7 +1,9 @@
-import { addMonths, isAfter, parseISO } from 'date-fns';
+import { addDays, addMonths, isAfter } from 'date-fns';
 
 export const verifyIfIsNewJoiner = (joinedAt: string, providerAccountCreatedAt?: string) => {
-  const dateToCompare = parseISO(providerAccountCreatedAt || joinedAt);
+  let dateToCompare = new Date(providerAccountCreatedAt || joinedAt);
+
+  dateToCompare = addDays(dateToCompare, 15);
 
   const maxDateToBeNewJoiner = addMonths(dateToCompare, 3);