+++ CHANGELOG.md
@@ -2,7 +2,58 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased](https://github.com/xgeekshq/split/compare/v0.1.6...HEAD)
+## [Unreleased](https://github.com/xgeekshq/split/compare/v0.1.10...HEAD)
+
+## [v0.1.10](https://github.com/xgeekshq/split/compare/v0.1.9...v0.1.10) - 2023-02-02
+
+### What Changed 👀
+
+### 🐛 Bug Fixes
+
+- fix: new responsible validation @nunocaseiro (#1001)
+- fix: isNewJoiner period and votes @CatiaAntunes96 (#1000)
+
+### 📄 Documentation
+
+- feat: popover primitive refactor and story @JoaoSaIvador (#993)
+
+**Full Changelog**: https://github.com/xgeekshq/split/compare/v0.1.9...v0.1.10
+
+## [v0.1.9](https://github.com/xgeekshq/split/compare/v0.1.8...v0.1.9) - 2023-02-02
+
+### What Changed 👀
+
+- docs: fix typo @dpompeu-xgeeks (#987)
+
+### 🐛 Bug Fixes
+
+- fix: slack channels @CatiaAntunes96 (#992)
+
+**Full Changelog**: https://github.com/xgeekshq/split/compare/v0.1.8...v0.1.9
+
+## [v0.1.8](https://github.com/xgeekshq/split/compare/v0.1.7...v0.1.8) - 2023-02-01
+
+### What Changed 👀
+
+### 🐛 Bug Fixes
+
+- fix: main board settings @CatiaAntunes96 (#983)
+
+### 📄 Documentation
+
+- feat: avatar primitive story @JoaoSaIvador (#980)
+
+**Full Changelog**: https://github.com/xgeekshq/split/compare/v0.1.7...v0.1.8
+
+## [v0.1.7](https://github.com/xgeekshq/split/compare/v0.1.6...v0.1.7) - 2023-01-31
+
+### What Changed 👀
+
+### 🐛 Bug Fixes
+
+- fix: participants @nunocaseiro (#978)
+
+**Full Changelog**: https://github.com/xgeekshq/split/compare/v0.1.6...v0.1.7
 
 ## [v0.1.6](https://github.com/xgeekshq/split/compare/v0.1.5...v0.1.6) - 2023-01-31